working waveshaping with test bench
